+++ prov/efa/docs/building.md
@@ -50,6 +50,6 @@ CFLAGS that might be useful:
 * `EFA_PERF_ENABLED`: enable the perf hooks to determine cycle/instruction count
 for functions in the send/receive/completion paths. See fi_hook(7) and the
 Linux perf documentation for more information.
-* `ENABLE_EFA_RDM_PKT_DUMP`: turn on packet dump prints, very verbose. These
+* `ENABLE_EFA_RDM_PKE_DUMP`: turn on packet dump prints, very verbose. These
 functions haven't been kept up to date with recent protocol changes so this
 might not be useful until fixed.
