+++ prov/efa/src/rdm/efa_rdm_pke_cmd.h
@@ -32,9 +32,5 @@ void efa_rdm_pke_proc_received(struct efa_rdm_pke *pkt_entry, struct efa_rdm_pee
 
 void efa_rdm_pke_proc_received_no_hdr(struct efa_rdm_pke *pkt_entry, bool has_imm_data, uint32_t imm_data);
 
-#if ENABLE_DEBUG
-void efa_rdm_pke_print(struct efa_rdm_pke *pkt_entry, char *prefix);
-#endif
-
 #endif
